Move CTFE handling of nondiverging intrinsics to intrinsics.rs
